Unify initialization process for both Uniform and DiagonalMass
Now the mass is initialized with same data name (from a user standpoint)
If user specifies the data at the vertex use it.
If not, if the user specifies the massDensity (Diagonal) use it
Else, or if any of above initialization above goes wrong, use the
totalMass which is the default case and the easiest to understand
from a user standpoint.